+++ src/parkapi_sources/converters/__init__.py
@@ -12,6 +12,7 @@ from .bb_parkhaus import BBParkhausPushConverter
 from .bfrk_bw import BfrkBwBikePushConverter, BfrkBwCarPushConverter
 from .bietigheim_bissingen import BietigheimBissingenPullConverter
 from .ellwangen import EllwangenPushConverter
+from .esslingen import EsslingenPushConverter
 from .freiburg import (
     FreiburgParkAndRideRealtimePullConverter,
     FreiburgParkAndRideStaticPullConverter,
@@ -19,13 +20,16 @@ from .freiburg import (
 )
 from .freiburg_disabled_sensors import FreiburgDisabledSensorsPullConverter
 from .freiburg_disabled_static import FreiburgDisabledStaticPullConverter
+from .freiburg_scanner import FreiburgScannerPullConverter
 from .friedrichshafen_sensors import FriedrichshafenSensorsPullConverter
 from .goldbeck import GoldbeckPushConverter
 from .heidelberg import HeidelbergPullConverter
+from .heidelberg_easypark import HeidelbergEasyParkPullConverter
 from .herrenberg import HerrenbergPullConverter
 from .herrenberg_bike import HerrenbergBikePullConverter
 from .huefner import HuefnerPushConverter
 from .karlsruhe import KarlsruheBikePullConverter, KarlsruhePullConverter
+from .keltern import KelternPushConverter
 from .kienzler import (
     KienzlerBikeAndRidePullConverter,
     KienzlerKarlsruhePullConverter,
@@ -49,6 +53,7 @@ from .pum_bw import PumBwPushConverter
 from .radvis_bw import RadvisBwPullConverter
 from .reutlingen import ReutlingenPushConverter
 from .reutlingen_bike import ReutlingenBikePushConverter
+from .reutlingen_disabled import ReutlingenDisabledPushConverter
 from .stuttgart import StuttgartPushConverter
 from .ulm import UlmPullConverter
 from .ulm_sensors import UlmSensorsPullConverter